General.cs split into sepereate files for each class.
